LayoutTests/platform/gtk/media/controls-drag-timebar-expected.txt

Test that drag the timebar thumb causes seeks.

RUN(video.autoplay = true)
RUN(video.src = 'content/test.mp4')
EVENT(playing)
EVENT(seeked)
Time: 2.1
EVENT(seeked)
Time: 2.7
EVENT(seeked)
Time: 3.2
END OF TEST
